feat(heights): widen app-layer heights to u64 with lossless SQLite mapping (RFC-019 S0)
App-layer block heights were i32 (with saturating casts at the engine
boundary, including one feeding validator_set); the engine is u64
end-to-end. Regenesis (RFC-019) keeps heights continuous across epochs,
making the app layer the binding integer ceiling — widen everything to
u64 now, while meshes are disposable.

- hopnet-common: height_to_db/height_from_db bit-cast pair, lossless
  over the full 64-bit range; ordering contract documented (valid below
  2^63). Height::{from_db,as_db} and all height columns route through
  it; this rusqlite has no u64 ToSql/FromSql, so the compiler enforces
  the discipline.
- replicated tx payloads (Metric, PlacementUpdate, SelfCheckFragments,
  TakeoutPayload) and wire types (DecidedFetch, txforward) widen —
  legal while meshes are disposable.
- typeshare: parser rejects u64, so client-facing height fields use
  serialized_as U64Height mapped per language (UInt64/number/Long).
- Swift FileProvider: 8-byte item versions, UInt64 sync anchors.
- deletes the dead legacy Block/BlockData (i32 heights) from
  src/consensus/types.rs.
- orchestrator tests: detect container runtime instead of assuming
  Docker — growth-added nodes get host-port bindings under podman
  (required to run the mesh suites on this machine).
- full-range roundtrip tests (context, store reopen at u64::MAX,
  validator activation near 2^63); snapshotter labels unchanged.

+++ apple/HopNetFileProvider/Sources/HopNetFileProviderExtension/HopNetApiClient.swift
@@ -228,7 +228,7 @@ public class HopNetApiClient {
     
     // MARK: - Changes
     
-    public func getChanges(parentPath: String? = nil, sinceHeight: Int32? = nil) async throws -> ChangesResponse {
+    public func getChanges(parentPath: String? = nil, sinceHeight: UInt64? = nil) async throws -> ChangesResponse {
         let urlString = "\(config.baseUrl)/api/integrations/fileprovider/changes"
         
         guard var urlComponents = URLComponents(string: urlString) else {
