+++ src/metrics/types.rs
@@ -10,7 +10,7 @@ pub struct Metric {
     pub rtt_variance: Option<f64>,
     pub rtt_jitter: Option<f64>,
     pub throughput: Option<i64>,
-    pub height: i32,                   // Consensus height when measurement was taken
+    pub height: u64,                   // Consensus height when measurement was taken
     pub available: bool,               // Whether target node was reachable
     pub storage_total_gb: Option<u32>, // Total storage capacity in GB
     pub storage_used_gb: Option<u32>,  // Used storage capacity in GB
